homepage/requirements.txt
# Flask Web Framework
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-SocketIO>=5.3.0

# Database
SQLAlchemy>=2.0.0

# Machine Learning & Data Processing
joblib>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# WebSocket
python-socketio>=5.9.0
python-socketio[client]>=5.9.0
eventlet>=0.33.0

# Google Gemini API (for AI features)
google-genai>=0.1.0
google-generativeai>=0.3.0

# Additional Utilities
Werkzeug>=3.0.0

